ghost roles need to be desynced now.
