+++ Patches/PlayerControlPatch.cs
@@ -1870,7 +1870,6 @@ public static class PlayerControlDiePatch
 [HarmonyPatch(typeof(PlayerControl), nameof(PlayerControl.RpcSetRole))]
 class PlayerControlSetRolePatch
 {
-    public static readonly Dictionary<byte, bool> DidSetGhost = [];
     private static readonly Dictionary<PlayerControl, RoleTypes> ghostRoles = [];
     public static bool Prefix(PlayerControl __instance, [HarmonyArgument(0)] ref RoleTypes roleType, [HarmonyArgument(1)] ref bool canOverrideRole)
     {
@@ -1884,8 +1883,6 @@ class PlayerControlSetRolePatch
         // Ghost assign
         if (roleType is RoleTypes.CrewmateGhost or RoleTypes.ImpostorGhost)
         {
-            if (DidSetGhost.TryGetValue(target.PlayerId, out var isGhost) && isGhost) // Prevent double assignment if player gets killed as a ghost
-                return false;
 
             try
             {
@@ -1924,7 +1921,9 @@ class PlayerControlSetRolePatch
             if (ghostRoles.All(kvp => kvp.Value == RoleTypes.GuardianAngel))
             {
                 roleType = RoleTypes.GuardianAngel;
-                return true;
+                __instance.RpcSetRoleDesync(RoleTypes.GuardianAngel, __instance.GetClientId());
+                GhostRoleAssign.CreateGAMessage(__instance);
+                return false;
             }
             // If all players see player as Crewmate Ghost
             else if (ghostRoles.All(kvp => kvp.Value == RoleTypes.CrewmateGhost))
@@ -1958,69 +1957,11 @@ class PlayerControlSetRolePatch
 
         try
         {
-            if (roleType == RoleTypes.GuardianAngel && !DidSetGhost.ContainsKey(__instance.PlayerId))
-            {
-                Utils.NotifyRoles(SpecifyTarget: __instance);
-                _ = new LateTask(() => {
-
-                    __instance.RpcResetAbilityCooldown();
-
-                    if (Options.SendRoleDescriptionFirstMeeting.GetBool())
-                    {
-                        var host = PlayerControl.LocalPlayer;
-                        var name = host.Data.PlayerName;
-                        var lp = __instance;
-                        var sb = new StringBuilder();
-                        var conf = new StringBuilder();
-                        var role = __instance.GetCustomRole();
-                        var rlHex = Utils.GetRoleColorCode(role);
-                        sb.Append(Utils.GetRoleTitle(role) + lp.GetRoleInfo(true));
-                        if (Options.CustomRoleSpawnChances.TryGetValue(role, out var opt))
-                            Utils.ShowChildrenSettings(Options.CustomRoleSpawnChances[role], ref conf);
-                        var cleared = conf.ToString();
-                        conf.Clear().Append($"<size={ChatCommands.Csize}>" + $"<color={rlHex}>{GetString(role.ToString())} {GetString("Settings:")}</color>\n" + cleared + "</size>");
-
-                        var writer = CustomRpcSender.Create("SendGhostRoleInfo", SendOption.None);
-                        writer.StartMessage(__instance.GetClientId());
-                        writer.StartRpc(host.NetId, (byte)RpcCalls.SetName)
-                            .Write(host.Data.NetId)
-                            .Write(Utils.ColorString(Utils.GetRoleColor(role), GetString("GhostTransformTitle")))
-                            .EndRpc();
-                        writer.StartRpc(host.NetId, (byte)RpcCalls.SendChat)
-                            .Write(sb.ToString())
-                            .EndRpc();
-                        writer.EndMessage();
-                        writer.SendMessage();
-
-                        var writer2 = CustomRpcSender.Create("SendGhostRoleConfig", SendOption.None);
-                        writer2.StartMessage(__instance.GetClientId());
-                        writer2.StartRpc(host.NetId, (byte)RpcCalls.SendChat)
-                            .Write(conf.ToString())
-                            .EndRpc();
-                        writer2.StartRpc(host.NetId, (byte)RpcCalls.SetName)
-                            .Write(host.Data.NetId)
-                            .Write(name)
-                            .EndRpc();
-                        writer2.EndMessage();
-                        writer2.SendMessage();
-
-                        // Utils.SendMessage(sb.ToString(), __instance.PlayerId, Utils.ColorString(Utils.GetRoleColor(role), GetString("GhostTransformTitle")));
-
-                    }
-
-                }, 0.1f, $"SetGuardianAngel for playerId: {__instance.PlayerId}");
-            }
+            
 
             if (__runOriginal)
             {
                 Logger.Info($" {__instance.GetRealName()} => {roleType}", "PlayerControl.RpcSetRole");
-
-                if (roleType is RoleTypes.CrewmateGhost or RoleTypes.ImpostorGhost or RoleTypes.GuardianAngel)
-                {
-                    if (!DidSetGhost.ContainsKey(__instance.PlayerId))
-                        DidSetGhost.Add(__instance.PlayerId, true);
-
-                }
             }
             _ = new LateTask(() => { Main.AllPlayerControls.Do(x => Logger.Info($"{x.GetRealName()}/{x.Data.Role.GetType().Name}/{x.Data.Role.Role}", "Test All Roletypes, and Roleclass")); }, 3f);
         }
