+++ Resources/Lang/en_US.json
@@ -2274,7 +2274,7 @@
   "Message.YTPlanNotice": "Note: The [YouTuber Plan] is enabled in this lobby, which means the Host can specify their role in the next game to make it easier to get content. If the Host abuses this feature, please exit the game or report it.\nCurrent Creator Credentials:",
   "Message.OnlyCanBeUsedByHost": "<color=#ff1919><size=150%><b>ERROR</b></size></color>\n\n<color=#ff1919>This command may only be used by the host.</color>",
   "Message.MaxPlayers": "Maximum players set to ",
-  "Message.GhostRoleInfo": "<size=90%><b>Ghost Role Info</b></size>\nHiya! A little about ghost roles...<size=75%>\n\nGhost roles drastically impact the game, so it's not recommended for smaller lobbies if you're unfamiliar. If not explicitly stated otherwise in the description, the Guard button is their ability button ;)\n\n<b>Spawning:</b>\nGhost-roles only spawn after death; the first x people from (team) to die get them.</size>\n\n<size=50%><b>PS: If your previous role didn't have tasks(e.g., sheriff), your tasks as a ghost-role aren't needed for task-win</b></size>",
+  "Message.GhostRoleInfo": "<size=90%><b>Ghost Role Info</b></size>\nHiya! A little about ghost roles...<size=75%>\n\nGhost roles drastically impact the game, so it's not recommended for smaller lobbies if you're unfamiliar. If not explicitly stated otherwise in the description, the Guard button is their ability button ;)\n\n<b>Spawning:</b>\nGhost-roles only spawn after death; the first x people from (team) to die get them.</size>\n\n<size=50%><b>PS: If your previous role didn't have tasks(e.g., sheriff), your tasks as a ghost-role aren't needed for task-win</b></size>\n\n<size=50%><b>NB: You will see a death-animation of yourself after every meeting to fix blackscreen</b></size>",
   "ApocalypseInfoTitle": "Neutral Apocalypse Info:",
   "Message.ApocalypseInfo": "Every role of the <#ff174f>Apocalypse</color> Team has their own objective to carry out in order to transform.\n<#2B0804>Transformed</color> <#ff174f>Apocalypse</color> members have a drastic change on the game and are immortal (except for being voted), but everyone will be notified that they have transformed.\n\nRoles: <#e5f6b4>Plaguebearer</color>, <#A675A1>Soul Collector</color>, <#bf9f7a>Baker</color>, <#cc0044>Berserker</color>\nTransformed: <#343136>Pestilence</color>, <#644661>Death</color>, <#83461c>Famine</color>, <#2B0804>War</color>\n\nApocalypse members can see eachother's roles and ability icons.\nLike Neutral Killers, Apocalypse members keep the game going as well, have fun!",
   "Message.MeCommandInfo": "<size=100%><b>Hi [{0}] {1} !</b></size>\n\n<size=80%><align=\"flush\"><color=#f34c50>friend-code</color> <color=#a6a832>Hash-Puid</color> <color=#6632a8>Type</color> <br>{2} {3} {4}<br><br><color=#7a0000>IsDev</color> <color=#349972>HasUp</color> <size=50%><color=#252dc4>/color-Bypass</color></size><br>{5} {6} {7} <br><br></align>",
